+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 版本更新 
 
 ## v0.2
+### 2021/06/17
+- add aspectFill mode to imgMode component
+
 ### 2021/06/16
 - add aspectFit mode to imgMode component